完善了删除角色时角色被占用的逻辑，初步完成菜单界面

+++ src/main/java/com/quxue/wedding_host/service/impl/RoleServiceImpl.java
@@ -26,6 +26,9 @@ public class RoleServiceImpl implements RoleService {
     @Transactional
     @Override
     public Integer delRole(Integer id) {
+        if (!roleMapper.queryUsageBeforeDel(id).isEmpty()) {
+            return -1;
+        }
         return roleMapper.deleteById(id);
     }
 
@@ -62,6 +65,11 @@ public class RoleServiceImpl implements RoleService {
     @Transactional
     @Override
     public Integer batchDelRole(List<Integer> roleIds) {
+        for (Integer roleId : roleIds) {
+            if (!roleMapper.queryUsageBeforeDel(roleId).isEmpty()) {
+                return -1;
+            }
+        }
         return roleMapper.deleteBatchIds(roleIds);
     }
